+++ test/src/common/utils/thumbs.spec.js
@@ -1,4 +1,4 @@
-import {getThumbSlicesUrl} from '../../../src/common/utils/thumbs'
+import {getThumbSlicesUrl} from '../../../../src/common/utils/thumbs'
 
 describe('getThumbSlicesUrl', function () {
   const fakeData = {
